requirements.txt
# requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
plotly  # <-- ADD THIS LINE
python-docx
pypdf2